Extract virt-handler.go from infra_test.go
